--- src/ModelSchema.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Backpack\CRUD;
-
-class ModelSchema extends TableSchema
-{
-    public function __construct($model)
-    {
-        [$connection, $table] = $model::getConnectionAndTable();
-        $this->schema = DatabaseSchema::getForTable($connection->getName(), $table);
-    }
-}
